Keep wai-route routing table only for swagger1.2 (part 1).
